This should allow multiple people to make against the same stack
The problem was that key_name was a CF template parameter, and
everyone has a unique key_name so for each different person who
runs make the CF will create a new ec2 instance because the
key_name changed :(

The "solution" is to not tie the key_name to a person (duh).
Instead this uses the stack_name-fxadev for the key_name. The
plus is that now anyone can update the stack without it triggering
an ec2 rebuild. The downside is that now there's a key for each
stack that is "variable", it changes every time someone new runs
make, but that shouldn't really matter because its only important
on runs that make changes to the CF template because the team keys
get uploaded by the team role.

Added bonus, my_vars.yml is gone!