fix(mailer): Remove the `locales`, depend on fxa-shared
Only two locales supported are "en_US" and "de", yet the
defaultLanguage was set to "en". I see a lot of auth-mailer restarts
on fxa-dev when sending verification reminders and am wondering
if this could be a cause.

Instead of using a local set of `locales`, use those from
fxa-shared, which should contain `en` as a language.