Update auth-server redirect domain (#356) r=@vladikoff
This PR updates the REDIRECT_DOMAIN env in fxaci to support In mozilla/fxa-content-server#5922. In that PR we use the redirectTo param to redirect back to the original settings panel after a user has verified their email address. The auth-server rejects requests because redirect domain is 127.0.0.1.

This was tested in https://github.com/mozilla/fxa-dev/tree/docker-vbudhram and the corresponding tests pass.