Add channelserver and pairsona branches (#398)
Connects to #395.

With this PR I'm able to run channelserver as part of the fxa-dev stack, and I can send some test messages between two firefox windows. Next step is to get the pairsona branches deployed and configured correctly, then it should be possible to do an end-to-end pairing demo.

We can't merge this until we get channelserver building a proper docker image, see mozilla-services/channelserver#18

I also had to switch to using a new application load balancer rather than the "classic" ELB, in order to get the required combination of TLS and websockets. It wasn't too painful in practice, we might like to consider what it would look like to migrate to that setup across all our stacks. Or we might like to just keep the websocket thing nice and isolated for now :-)

Anyway, pushing this for early visibility, I'll keep poking at it this week.